backend/requirements.txt
fastapi
uvicorn
python-multipart
numpy
opencv-python-headless
pillow
rembg
onnxruntime
